moved MainSlider.tsx to Slider folder
