+++ src/app/components/Slider/MainSlider.tsx
@@ -1,9 +1,9 @@
 import React from 'react';
 import Slider from 'rc-slider';
 import Tooltip from 'rc-tooltip';
-import {useAppSelector, useAppDispatch} from '../state-management/hooks';
-import {setRenderIndex} from '../state-management/slices/SnapshotSlice';
-import {selectFilterState} from '../state-management/slices/FilterSlice';
+import {useAppSelector, useAppDispatch} from '../../state-management/hooks';
+import {setRenderIndex} from '../../state-management/slices/SnapshotSlice';
+import {selectFilterState} from '../../state-management/slices/FilterSlice';
 
 const {Handle} = Slider;
