+++ src/bev-generated-stats.json
@@ -1,0 +1,32 @@
+{
+  "errors": [
+    "Entry module not found: Error: Can't resolve './src/app/index.tsx' in '/Users/thetinyego/Desktop/CodeSmith/Recoilize/src'\nresolve './src/app/index.tsx' in '/Users/thetinyego/Desktop/CodeSmith/Recoilize/src'\n  using description file: /Users/thetinyego/Desktop/CodeSmith/Recoilize/package.json (relative path: ./src)\n    Field 'browser' doesn't contain a valid alias configuration\n    using description file: /Users/thetinyego/Desktop/CodeSmith/Recoilize/package.json (relative path: ./src/src/app/index.tsx)\n      no extension\n        Field 'browser' doesn't contain a valid alias configuration\n        /Users/thetinyego/Desktop/CodeSmith/Recoilize/src/src/app/index.tsx doesn't exist\n      .tsx\n        Field 'browser' doesn't contain a valid alias configuration\n        /Users/thetinyego/Desktop/CodeSmith/Recoilize/src/src/app/index.tsx.tsx doesn't exist\n      .ts\n        Field 'browser' doesn't contain a valid alias configuration\n        /Users/thetinyego/Desktop/CodeSmith/Recoilize/src/src/app/index.tsx.ts doesn't exist\n      .js\n        Field 'browser' doesn't contain a valid alias configuration\n        /Users/thetinyego/Desktop/CodeSmith/Recoilize/src/src/app/index.tsx.js doesn't exist\n      .jsx\n        Field 'browser' doesn't contain a valid alias configuration\n        /Users/thetinyego/Desktop/CodeSmith/Recoilize/src/src/app/index.tsx.jsx doesn't exist\n      as directory\n        /Users/thetinyego/Desktop/CodeSmith/Recoilize/src/src/app/index.tsx doesn't exist",
+    "Entry module not found: Error: Can't resolve './src/extension/background.ts' in '/Users/thetinyego/Desktop/CodeSmith/Recoilize/src'\nresolve './src/extension/background.ts' in '/Users/thetinyego/Desktop/CodeSmith/Recoilize/src'\n  using description file: /Users/thetinyego/Desktop/CodeSmith/Recoilize/package.json (relative path: ./src)\n    Field 'browser' doesn't contain a valid alias configuration\n    using description file: /Users/thetinyego/Desktop/CodeSmith/Recoilize/package.json (relative path: ./src/src/extension/background.ts)\n      no extension\n        Field 'browser' doesn't contain a valid alias configuration\n        /Users/thetinyego/Desktop/CodeSmith/Recoilize/src/src/extension/background.ts doesn't exist\n      .tsx\n        Field 'browser' doesn't contain a valid alias configuration\n        /Users/thetinyego/Desktop/CodeSmith/Recoilize/src/src/extension/background.ts.tsx doesn't exist\n      .ts\n        Field 'browser' doesn't contain a valid alias configuration\n        /Users/thetinyego/Desktop/CodeSmith/Recoilize/src/src/extension/background.ts.ts doesn't exist\n      .js\n        Field 'browser' doesn't contain a valid alias configuration\n        /Users/thetinyego/Desktop/CodeSmith/Recoilize/src/src/extension/background.ts.js doesn't exist\n      .jsx\n        Field 'browser' doesn't contain a valid alias configuration\n        /Users/thetinyego/Desktop/CodeSmith/Recoilize/src/src/extension/background.ts.jsx doesn't exist\n      as directory\n        /Users/thetinyego/Desktop/CodeSmith/Recoilize/src/src/extension/background.ts doesn't exist",
+    "Entry module not found: Error: Can't resolve './src/extension/contentScript.ts' in '/Users/thetinyego/Desktop/CodeSmith/Recoilize/src'\nresolve './src/extension/contentScript.ts' in '/Users/thetinyego/Desktop/CodeSmith/Recoilize/src'\n  using description file: /Users/thetinyego/Desktop/CodeSmith/Recoilize/package.json (relative path: ./src)\n    Field 'browser' doesn't contain a valid alias configuration\n    using description file: /Users/thetinyego/Desktop/CodeSmith/Recoilize/package.json (relative path: ./src/src/extension/contentScript.ts)\n      no extension\n        Field 'browser' doesn't contain a valid alias configuration\n        /Users/thetinyego/Desktop/CodeSmith/Recoilize/src/src/extension/contentScript.ts doesn't exist\n      .tsx\n        Field 'browser' doesn't contain a valid alias configuration\n        /Users/thetinyego/Desktop/CodeSmith/Recoilize/src/src/extension/contentScript.ts.tsx doesn't exist\n      .ts\n        Field 'browser' doesn't contain a valid alias configuration\n        /Users/thetinyego/Desktop/CodeSmith/Recoilize/src/src/extension/contentScript.ts.ts doesn't exist\n      .js\n        Field 'browser' doesn't contain a valid alias configuration\n        /Users/thetinyego/Desktop/CodeSmith/Recoilize/src/src/extension/contentScript.ts.js doesn't exist\n      .jsx\n        Field 'browser' doesn't contain a valid alias configuration\n        /Users/thetinyego/Desktop/CodeSmith/Recoilize/src/src/extension/contentScript.ts.jsx doesn't exist\n      as directory\n        /Users/thetinyego/Desktop/CodeSmith/Recoilize/src/src/extension/contentScript.ts doesn't exist"
+  ],
+  "warnings": [
+    "configuration\nThe 'mode' option has not been set, webpack will fallback to 'production' for this value. Set 'mode' option to 'development' or 'production' to enable defaults for each environment.\nYou can also set it to 'none' to disable any default behavior. Learn more: https://webpack.js.org/configuration/mode/"
+  ],
+  "version": "4.46.0",
+  "hash": "8b59224c3715bdf86e72",
+  "time": 37,
+  "builtAt": 1642102300275,
+  "publicPath": "",
+  "outputPath": "/Users/thetinyego/Desktop/CodeSmith/Recoilize/src/extension/build/bundles",
+  "assetsByChunkName": {},
+  "assets": [],
+  "filteredAssets": 0,
+  "entrypoints": {},
+  "namedChunkGroups": {},
+  "chunks": [],
+  "modules": [],
+  "filteredModules": 0,
+  "logging": {
+    "webpack.buildChunkGraph.visitModules": {
+      "entries": [],
+      "filteredEntries": 1,
+      "debug": false
+    }
+  },
+  "children": []
+}
